+++ db/migrations/20200416093003_create_log_sorted_offer.sql
@@ -4,7 +4,7 @@ CREATE TABLE maker.log_sorted_offer
     id         SERIAL PRIMARY KEY,
     log_id     BIGINT  NOT NULL REFERENCES public.event_logs (id) ON DELETE CASCADE,
     header_id  INTEGER NOT NULL REFERENCES public.headers (id) ON DELETE CASCADE,
-    address_id INTEGER NOT NULL REFERENCES public.addresses (id) ON DELETE CASCADE,
+    address_id BIGINT  NOT NULL REFERENCES public.addresses (id) ON DELETE CASCADE,
     offer_id   NUMERIC,
     UNIQUE (header_id, log_id)
 );
